functions/template.go
package functions

import (
	"fmt"
	"strings"
	"text/template"
)

// TemplateFunctions operate on templates themselves
func TemplateFunctions(t *template.Template) template.FuncMap {
	return template.FuncMap{
		"include": GenerateIncludeFn(t),
	}
}

// GenerateIncludeFn creates a function to be used as an "include" function in templates
func GenerateIncludeFn(t *template.Template) func(string, interface{}) (string, error) {
	const recursionLimit = 100
	inc := make(map[string]int) // keep track of how many times each template has been nested
	return func(name string, data interface{}) (string, error) {
		var buf strings.Builder
		if inc[name] > recursionLimit {
			return "", fmt.Errorf("recursion limit hit rendering template: %s", name)
		}
		inc[name]++
		err := t.ExecuteTemplate(&buf, name, data)
		inc[name]--
		return buf.String(), err
	}
}
